glue_python/aliyuntvtoken/requirements.txt
flask
requests
pycryptodome